refactored BoilerStateHandler
